refactor: method에서 특화된 네이밍으로 변경
